added inventory screen

+++ app/src/main/java/com/example/explory/data/model/inventory/CosmeticItemInInventoryDto.kt
@@ -4,7 +4,7 @@ import com.example.explory.data.model.shop.CosmeticType
 import com.example.explory.data.model.shop.RarityType
 
 data class CosmeticItemInInventoryDto(
-    val itemId: Int,
+    val itemId: Long,
     val name: String,
     val description: String,
     val price: Int,
